Refactor code and add sprign security

+++ src/main/java/com/acelerati/management_service/application/dto/response/ProductDTO.java
@@ -3,16 +3,15 @@ package com.acelerati.management_service.application.dto.response;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 public class ProductDTO {
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
     private String model;
-    private  Integer idBrand;
-    private Integer idCategory;
+    private  Long idBrand;
+    private Long idCategory;
 }
